pyproject.toml
[project]
name = "aegis-detect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=0.20.0",
    "datasets>=2.0.0",
    "huggingface-hub>=0.15.0",
    "matplotlib>=3.10.7",
    "numpy>=1.21.0",
    "openai>=1.0.0",
    "pandas>=1.5.0",
    "peft>=0.4.0",
    "python-dotenv>=0.19.0",
    "safetensors>=0.3.0",
    "scikit-learn>=1.0.0",
    "seaborn>=0.13.2",
    "torch>=2.0.0",
    "tqdm>=4.60.0",
    "transformers>=4.30.0",
    "ruff>=0.6.0",
]

[project.scripts]
aegis = "aegis.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["aegis"]

[tool.ruff]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "node_modules",
    "build",
    "dist",
    "*.egg-info",
    "saved_models",
    "data/raw",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
